python/requirements.txt
# Python dependencies for Clash of Clans Bot Handlers
# Основные зависимости для работы с Telegram Bot API
python-telegram-bot>=20.0

# Для работы с базой данных (если используется SQLite)
aiosqlite>=0.19.0

# Для HTTP запросов к Clash of Clans API
aiohttp>=3.8.0

# Для работы с JSON
# json встроен в Python

# Для логирования
# logging встроен в Python

# Для асинхронного программирования
# asyncio встроен в Python

# Для типизации
# typing встроен в Python

# Дополнительные зависимости (опционально)
# Для более продвинутого логирования
structlog>=23.0.0

# Для валидации данных
pydantic>=2.0.0

# Для работы с датами
python-dateutil>=2.8.0